Create IObjectSnapshot interface in core
ObjectSnapshot relies on too many things from the DB to be moved to core
easily, so we move it back to SIL.Harmony.Db and create an interface in
core instead. Then the only difference is that the type of Commit will
be different depending on whether you access it via the interface (in
which case you'll get a CommitBase from core) or via the implementation
class (in which case you'll get a Commit object from SIL.Harmony).
